+++ test/core/ActionTest.cxx
@@ -63,7 +63,7 @@ protected:
 		beaconConfiguration = std::make_shared<configuration::BeaconConfiguration>();
 		beaconCacheConfiguration = std::make_shared<configuration::BeaconCacheConfiguration>(-1, -1, -1);
 		configuration = std::shared_ptr<configuration::Configuration>(new configuration::Configuration(device, configuration::OpenKitType::Type::DYNATRACE,
-			core::UTF8String(APP_NAME), "", APP_ID, 0, "",
+			core::UTF8String(APP_NAME), "", APP_ID, 0, "0", "",
 			sessionIDProvider, trustManager, beaconCacheConfiguration, beaconConfiguration));
 		configuration->enableCapture();